Updated to use FFMPEG version 6 libraries through reference to FFmpeg.AutoGen package. Also included these updates: - Change some test programs to run under .NET 6.0 - Update README to indicate that FFMPEG binaries are not included with Windows - Look in system PATH for the FFMPEG binaries - Throw exception if binary path not found by any method

+++ src/FFmpegAudioDecoder.cs
@@ -137,7 +137,7 @@ namespace SIPSorceryMedia.FFmpeg
                 }
                 else
                     ffmpeg.av_opt_set_channel_layout(_swrContext, "in_channel_layout", (long)_audDecCtx->channel_layout, 0);
-                ffmpeg.av_opt_set_channel_layout(_swrContext, "out_channel_layout", ffmpeg.AV_CH_LAYOUT_MONO, 0);
+                ffmpeg.av_opt_set_channel_layout(_swrContext, "out_channel_layout", (long)ffmpeg.AV_CH_LAYOUT_MONO, 0);
 
                 if ( ffmpeg.swr_init(_swrContext) < 0 )
                 {
